internal: refactor missing unsafe diagnostic
